# docs/openapi.json
{"openapi":"3.1.0","servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Bearer Authentication":[]}],"tags":[{"name":"Admin - Cards","description":"Card management for administrators"},{"name":"Cards","description":"Self-service endpoints for a user's own cards"},{"name":"Authentication","description":"Registration and JWT login"}],"paths":{"/api/cards/{cardId}/block-request":{"post":{"tags":["Cards"],"operationId":"requestBlock","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}}},"/api/cards/transfer":{"post":{"tags":["Cards"],"operationId":"transfer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransferRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransferResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/auth/refresh":{"post":{"tags":["Authentication"],"operationId":"refresh","parameters":[{"name":"X-Refresh-Token","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["Authentication"],"operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/admin/cards":{"get":{"tags":["Admin - Cards"],"operationId":"getAllCards","parameters":[{"name":"ownerId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","BLOCKED","EXPIRED"]}},{"name":"last4","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["createdAt,ASC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseCardResponse"}}}}}},"post":{"tags":["Admin - Cards"],"operationId":"createCard","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCardRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}}},"/api/admin/cards/{cardId}/block":{"post":{"tags":["Admin - Cards"],"operationId":"blockCard","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}}},"/api/admin/cards/{cardId}/activate":{"post":{"tags":["Admin - Cards"],"operationId":"activateCard","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}}},"/api/cards":{"get":{"tags":["Cards"],"operationId":"getMyCards","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","BLOCKED","EXPIRED"]}},{"name":"last4","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"type":"integer","default":0,"minimum":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"type":"integer","default":20,"minimum":1}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","default":["createdAt,ASC"],"items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResponseCardResponse"}}}}}}},"/api/cards/{cardId}":{"get":{"tags":["Cards"],"operationId":"getMyCard","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}}},"/api/cards/{cardId}/balance":{"get":{"tags":["Cards"],"operationId":"getBalance","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"number"}}}}}}}},"/api/admin/cards/{cardId}":{"get":{"tags":["Admin - Cards"],"operationId":"getCard","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CardResponse"}}}}}},"delete":{"tags":["Admin - Cards"],"operationId":"deleteCard","parameters":[{"name":"cardId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"CardResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"maskedNumber":{"type":"string"},"ownerId":{"type":"integer","format":"int64"},"ownerFullName":{"type":"string"},"expirationDate":{"type":"string","format":"date"},"status":{"type":"string","enum":["ACTIVE","BLOCKED","EXPIRED"]},"balance":{"type":"number"},"createdAt":{"type":"string","format":"date-time"}}},"TransferRequest":{"type":"object","properties":{"fromCardId":{"type":"integer","format":"int64"},"toCardId":{"type":"integer","format":"int64"},"amount":{"type":"number","minimum":0.01}},"required":["amount","fromCardId","toCardId"]},"TransferResponse":{"type":"object","properties":{"fromCardId":{"type":"integer","format":"int64"},"toCardId":{"type":"integer","format":"int64"},"amount":{"type":"number"},"fromCardNewBalance":{"type":"number"},"toCardNewBalance":{"type":"number"}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"fullName":{"type":"string","maxLength":100,"minLength":2},"phoneNumber":{"type":"string","minLength":1,"pattern":"^\\+?[1-9]\\d{1,14}$"},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","fullName","password","phoneNumber"]},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"},"expiresIn":{"type":"integer","format":"int64"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"CreateCardRequest":{"type":"object","properties":{"ownerId":{"type":"integer","format":"int64"},"expirationDate":{"type":"string","format":"date"}},"required":["expirationDate","ownerId"]},"PageResponseCardResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/CardResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"last":{"type":"boolean"}}}},"securitySchemes":{"Bearer Authentication":{"type":"http","description":"Paste only the JWT access token (no 'Bearer ' prefix needed)","scheme":"bearer","bearerFormat":"JWT"}}}}